backend/requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=4.8.0
python-dotenv>=0.19.0
Pillow>=8.3.1
django-cors-headers>=3.8.0
stripe>=3.1.0
